backend/requirements.txt
Flask
Flask-Cors
google-generativeai
python-docx
PyPDF2